Rewrite Meter_humanUnit() to accept floating point value
Since Meter_humanUnit() is often called with floating point values in
Meter objects, rewrite the function to let it process `double` type
natively, and save floating point to integer casts.

The rewritten function:
* Allows higher orders of magnitude including 'R' and 'Q', and
  addresses infinity. (The previous version has a maximum value of
  (2^64 - 1) representing 16 ZiB.)
* Rounds values when they are in intervals (99.9, 100) and (9.99, 10),
  and displays them with correct precision (number of fraction digits).
* Produces assertion error on negative and NaN values (undefined
  behavior).

+++ pcp/PCPDynamicMeter.c
@@ -351,27 +351,27 @@ void PCPDynamicMeter_updateValues(PCPDynamicMeter* this, Meter* meter) {
             break;
          case PM_TYPE_32:
             bytes += conv.dimSpace ?
-               Meter_humanUnit(buffer + bytes, atom.l, size - bytes) :
+               Meter_humanUnit(buffer + bytes, (double) atom.l, size - bytes) :
                xSnprintf(buffer + bytes, size - bytes, "%d", atom.l);
             break;
          case PM_TYPE_U32:
             bytes += conv.dimSpace ?
-               Meter_humanUnit(buffer + bytes, atom.ul, size - bytes) :
+               Meter_humanUnit(buffer + bytes, (double) atom.ul, size - bytes) :
                xSnprintf(buffer + bytes, size - bytes, "%u", atom.ul);
             break;
          case PM_TYPE_64:
             bytes += conv.dimSpace ?
-               Meter_humanUnit(buffer + bytes, atom.ll, size - bytes) :
+               Meter_humanUnit(buffer + bytes, (double) atom.ll, size - bytes) :
                xSnprintf(buffer + bytes, size - bytes, "%lld", (long long) atom.ll);
             break;
          case PM_TYPE_U64:
             bytes += conv.dimSpace ?
-               Meter_humanUnit(buffer + bytes, atom.ull, size - bytes) :
+               Meter_humanUnit(buffer + bytes, (double) atom.ull, size - bytes) :
                xSnprintf(buffer + bytes, size - bytes, "%llu", (unsigned long long) atom.ull);
             break;
          case PM_TYPE_FLOAT:
             bytes += conv.dimSpace ?
-               Meter_humanUnit(buffer + bytes, atom.f, size - bytes) :
+               Meter_humanUnit(buffer + bytes, (double) atom.f, size - bytes) :
                xSnprintf(buffer + bytes, size - bytes, "%.2f", (double) atom.f);
             break;
          case PM_TYPE_DOUBLE:
@@ -427,27 +427,27 @@ void PCPDynamicMeter_display(PCPDynamicMeter* this, ATTR_UNUSED const Meter* met
             break;
          case PM_TYPE_32:
             len = conv.dimSpace ?
-               Meter_humanUnit(buffer, atom.l, sizeof(buffer)) :
+               Meter_humanUnit(buffer, (double) atom.l, sizeof(buffer)) :
                xSnprintf(buffer, sizeof(buffer), "%d", atom.l);
             break;
          case PM_TYPE_U32:
             len = conv.dimSpace ?
-               Meter_humanUnit(buffer, atom.ul, sizeof(buffer)) :
+               Meter_humanUnit(buffer, (double) atom.ul, sizeof(buffer)) :
                xSnprintf(buffer, sizeof(buffer), "%u", atom.ul);
             break;
          case PM_TYPE_64:
             len = conv.dimSpace ?
-               Meter_humanUnit(buffer, atom.ll, sizeof(buffer)) :
+               Meter_humanUnit(buffer, (double) atom.ll, sizeof(buffer)) :
                xSnprintf(buffer, sizeof(buffer), "%lld", (long long) atom.ll);
             break;
          case PM_TYPE_U64:
             len = conv.dimSpace ?
-               Meter_humanUnit(buffer, atom.ull, sizeof(buffer)) :
+               Meter_humanUnit(buffer, (double) atom.ull, sizeof(buffer)) :
                xSnprintf(buffer, sizeof(buffer), "%llu", (unsigned long long) atom.ull);
             break;
          case PM_TYPE_FLOAT:
             len = conv.dimSpace ?
-               Meter_humanUnit(buffer, atom.f, sizeof(buffer)) :
+               Meter_humanUnit(buffer, (double) atom.f, sizeof(buffer)) :
                xSnprintf(buffer, sizeof(buffer), "%.2f", (double) atom.f);
             break;
          case PM_TYPE_DOUBLE:
